Rename feature-scene-listing-button
